+++ test/crow/config_test.exs
@@ -1,0 +1,5 @@
+defmodule Crow.ConfigTest do
+  use ExUnit.Case, async: true
+
+  doctest Crow.Config
+end